server/requirements.txt
Flask==2.2
Flask-SQLAlchemy==3.0.2
python-dotenv==0.20.0
pytest==7.1.1
black==22.3.0
